requirements.txt
click>=8.1
openai>=1.25
python-dotenv>=1.0
rich>=13.7
